tests repackaged into classes based on the method, e.g. create, retrieve, upload, search, etc.
